.github/actions/cosmosdb-restore/requirements.txt
azure-cosmos==4.9.0
azure-storage-blob==12.25.1
azure-mgmt-storage==21.2.0
azure-identity==1.21.0